cleanup
-redudnant code
-treat warnings as errros
-add config await
