+++ src/SampleEndpoint/SomeCommand.cs
@@ -1,0 +1,6 @@
+﻿using NServiceBus;
+
+class SomeCommand : ICommand
+{
+    
+}
