+++ src/SampleEndpoint/SomeCommandHandler.cs
@@ -1,0 +1,7 @@
+﻿using System.Threading.Tasks;
+using NServiceBus;
+
+class SomeCommandHandler : IHandleMessages<SomeCommand>
+{
+    public Task Handle(SomeCommand message, IMessageHandlerContext context) => Task.CompletedTask;
+}
